mymaxpooling - back to float32

+++ src/cnn_lib.py
@@ -893,8 +893,8 @@ class MyMaxPooling(Layer):
             inputs, ksize=ksize, strides=self.strides,
             padding=self.padding.upper(), include_batch_in_index=True)
 
-        # argmax = tf.cast(argmax, tf.int32, name='cast_maxpooling')
-        argmax = tf.cast(argmax, tf.float32, name='cast_maxpooling')
+        argmax = tf.cast(argmax, tf.int32, name='cast_maxpooling')
+        # argmax = tf.cast(argmax, tf.float32, name='cast_maxpooling')
 
         return output, argmax
